// src/utils/types.ts
export type commonDIR3Unit = {
  __EMPTY: string;
  C_ID_UD_ORGANICA: string;
  C_DNM_UD_ORGANICA: string;
  C_ID_NIVEL_ADMON: number;
  C_ID_TIPO_ENT_PUBLICA: string;
  N_NIVEL_JERARQUICO: number;
  C_ID_DEP_UD_SUPERIOR: string;
  C_DNM_UD_ORGANICA_SUPERIOR: string;
  C_ID_DEP_UD_PRINCIPAL: string;
  C_DNM_UD_ORGANICA_PRINCIPAL: string;
  B_SW_DEP_EDP_PRINCIPAL: string;
  C_ID_DEP_EDP_PRINCIPAL: string;
  C_DNM_UD_ORGANICA_EDP_PRINCIPAL: string;
  C_ID_ESTADO: string;
  D_VIG_ALTA_OFICIAL: string;
  NIF_CIF: string;
  C_DESC_PROV: number;
  C_ID_AMB_PROVINCIA: string;
};